Fixed some sharing issues, and row returning issues.
Sharing issues pertain to complex sharing $_filters and the like with it's sub-classes. Added in row filtering so that a certain array form would be returned, upon a select statement, and that array form would be dependent upon a flag dictating that form. Default is set to filter results, but other options include returning the full monty raw, or still filtering it (slightly) but returning everything as an array, rather than a string (as is potential when it comes to getting only one cell returned on a very specific query).